Fixed AccessViolationException!
/ Fixed AccessViolationException on AntiDebugLibNative.cs (was just a typo)
/ Fixed check methods are returning true by default (thus causing whole false positives)
/ Fixed SandboxieCrasher to catch AccessViolationException (added attribute)
/ Fixed LoadLibraryA check failed to find word.exe (was actually 'write.exe' XD)
/ Fixed OutputDebugString comparison is inverted

+++ AntiDebugLib/Native/Kernel32+Indirect.cs
@@ -65,7 +65,7 @@ namespace AntiDebugLib.Native
         internal delegate bool DCloseHandle(IntPtr handle);
 
         [UnmanagedFunctionPointer(CallingConvention.Winapi)]
-        internal delegate bool DIsWow64Process2(IntPtr process, out uint processMachine, out uint nativeMatchine);
+        internal delegate bool DIsWow64Process2(IntPtr process, out ushort processMachine, out ushort nativeMatchine);
 
         #endregion
